+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/Constant.java
@@ -68,64 +68,54 @@ public class Constant extends TensorGenerator {
   }
 
   @Override
-  protected EnumSet<DType> getDTypes(PropagationCallGraphBuilder builder) {
+  protected EnumSet<DType> getDefaultDTypes(PropagationCallGraphBuilder builder) {
     EnumSet<DType> ret = EnumSet.noneOf(DType.class);
     PointerAnalysis<InstanceKey> pointerAnalysis = builder.getPointerAnalysis();
 
-    // The dtype is the second explicit argument.
-    // FIXME: Handle keyword arguments.
-    PointerKey dTypePointerKey = pointerAnalysis.getHeapModel().getPointerKeyForLocal(node, 3);
-    OrdinalSet<InstanceKey> dTypePointsToSet = pointerAnalysis.getPointsToSet(dTypePointerKey);
+    // If the argument dtype is not specified, then the type is inferred from the type of value.
+    // TODO: Handle keyword arguments.
+    PointerKey valuePK = pointerAnalysis.getHeapModel().getPointerKeyForLocal(node, 2);
 
-    // If the argument dtype is not specified,
-    if (dTypePointsToSet.isEmpty()) {
-      // then the type is inferred from the type of value.
-      PointerKey valuePK = pointerAnalysis.getHeapModel().getPointerKeyForLocal(node, 2);
+    for (InstanceKey valueIK : pointerAnalysis.getPointsToSet(valuePK))
+      if (valueIK instanceof ConstantKey) { // It's a scalar value.
+        ConstantKey<?> constantKey = (ConstantKey<?>) valueIK;
+        Object value = constantKey.getValue();
 
-      for (InstanceKey valueIK : pointerAnalysis.getPointsToSet(valuePK))
-        if (valueIK instanceof ConstantKey) { // It's a scalar value.
-          ConstantKey<?> constantKey = (ConstantKey<?>) valueIK;
-          Object value = constantKey.getValue();
-
-          if (value instanceof Float || value instanceof Double) {
-            ret.add(FLOAT32);
-            LOGGER.info(
-                "Inferred dtype: "
-                    + FLOAT32
-                    + " for source: "
-                    + source
-                    + " from value: "
-                    + value
-                    + ".");
-          } else if (value instanceof Integer || value instanceof Long) {
-            ret.add(INT32);
-            LOGGER.info(
-                "Inferred dtype: "
-                    + INT32
-                    + " for source: "
-                    + source
-                    + " from value: "
-                    + value
-                    + ".");
-          } else if (value instanceof String) {
-            ret.add(STRING);
-            LOGGER.info(
-                "Inferred dtype: "
-                    + STRING
-                    + " for source: "
-                    + source
-                    + " from value: "
-                    + value
-                    + ".");
-          } else
-            throw new IllegalStateException("Unknown constant type: " + value.getClass() + ".");
-        } else // TODO: More cases.
+        if (value instanceof Float || value instanceof Double) {
+          ret.add(FLOAT32);
+          LOGGER.info(
+              "Inferred dtype: "
+                  + FLOAT32
+                  + " for source: "
+                  + source
+                  + " from value: "
+                  + value
+                  + ".");
+        } else if (value instanceof Integer || value instanceof Long) {
+          ret.add(INT32);
+          LOGGER.info(
+              "Inferred dtype: "
+                  + INT32
+                  + " for source: "
+                  + source
+                  + " from value: "
+                  + value
+                  + ".");
+        } else if (value instanceof String) {
+          ret.add(STRING);
+          LOGGER.info(
+              "Inferred dtype: "
+                  + STRING
+                  + " for source: "
+                  + source
+                  + " from value: "
+                  + value
+                  + ".");
+        } else throw new IllegalStateException("Unknown constant type: " + value.getClass() + ".");
+      } else
+        // TODO: More cases.
         throw new IllegalStateException(
-              "Expected a " + ConstantKey.class + " for value, but got: " + valueIK + ".");
-    } else
-      // The dtype points-to set is non-empty, meaning that the dtype was explicitly set. TODO:
-      // Handle explicit dtypes.
-      throw new IllegalStateException("Explicit dtype set: " + dTypePointsToSet + ".");
+            "Expected a " + ConstantKey.class + " for value, but got: " + valueIK + ".");
 
     return ret;
   }
